fix: Tighten stx publishHook logic
