+++ app/scripts/lib/smart-transaction/smart-transactions.ts
@@ -24,7 +24,6 @@ import {
 } from '@metamask/transaction-controller';
 import type { Hex } from '@metamask/utils';
 import log from 'loglevel';
-
 import {
   ORIGIN_METAMASK,
   SMART_TRANSACTION_CONFIRMATION_TYPES,
@@ -38,10 +37,6 @@ import {
 } from '../../../../shared/modules/selectors';
 import { isLegacyTransaction } from '../../../../shared/modules/transaction.utils';
 import { ControllerFlatState } from '../../controller-init/controller-list';
-import { TransactionControllerInitMessenger } from '../../controller-init/messengers/transaction-controller-messenger';
-import { Delegation7702PublishHook } from '../transaction/hooks/delegation-7702-publish';
-import { isSendBundleSupported } from '../transaction/sentinel-api';
-import { getTransactionById } from '../transaction/util';
 
 const namespace = 'SmartTransactions';
 
@@ -546,7 +541,7 @@ function getUIState(flatState: ControllerFlatState) {
   return { metamask: flatState };
 }
 
-function getSmartTransactionCommonParams(
+export function getSmartTransactionCommonParams(
   flatState: ControllerFlatState,
   chainId?: string,
 ) {
@@ -567,108 +562,4 @@ function getSmartTransactionCommonParams(
     featureFlags,
     isHardwareWalletAccount,
   };
-}
-
-export async function publishHook({
-  flatState,
-  initMessenger,
-  signedTx,
-  smartTransactionsController,
-  transactionController,
-  transactionMeta,
-}: {
-  flatState: ControllerFlatState;
-  initMessenger: TransactionControllerInitMessenger;
-  signedTx: string;
-  smartTransactionsController: SmartTransactionsController;
-  transactionController: TransactionController;
-  transactionMeta: TransactionMeta;
-}) {
-  const { isSmartTransaction, featureFlags, isHardwareWalletAccount } =
-    getSmartTransactionCommonParams(flatState, transactionMeta.chainId);
-  const sendBundleSupport = await isSendBundleSupported(
-    transactionMeta.chainId,
-  );
-
-  if (!isSmartTransaction || !sendBundleSupport) {
-    const hook = new Delegation7702PublishHook({
-      isAtomicBatchSupported: transactionController.isAtomicBatchSupported.bind(
-        transactionController,
-      ),
-      messenger: initMessenger,
-    }).getHook();
-
-    return await hook(transactionMeta, signedTx);
-  }
-
-  if (
-    isSmartTransaction &&
-    (sendBundleSupport || transactionMeta.selectedGasFeeToken !== undefined)
-  ) {
-    const result = await submitSmartTransactionHook({
-      transactionMeta,
-      signedTransactionInHex: signedTx as Hex,
-      transactionController,
-      smartTransactionsController,
-      controllerMessenger: initMessenger,
-      isSmartTransaction,
-      isHardwareWallet: isHardwareWalletAccount,
-      // @ts-expect-error Smart transaction selector return type does not match FeatureFlags type from hook
-      featureFlags,
-    });
-
-    if (result?.transactionHash) {
-      return result;
-    }
-  }
-
-  // Default: fall back to regular transaction submission
-  return { transactionHash: undefined };
-}
-
-export function publishBatchHook({
-  transactionController,
-  smartTransactionsController,
-  hookControllerMessenger,
-  flatState,
-  transactions,
-}: {
-  transactionController: TransactionController;
-  smartTransactionsController: SmartTransactionsController;
-  hookControllerMessenger: SmartTransactionHookMessenger;
-  flatState: ControllerFlatState;
-  transactions: PublishBatchHookTransaction[];
-}) {
-  // Get transactionMeta based on the last transaction ID
-  const lastTransaction = transactions[transactions.length - 1];
-  const transactionMeta = getTransactionById(
-    lastTransaction.id ?? '',
-    transactionController,
-  );
-
-  // If we couldn't find the transaction, we should handle that gracefully
-  if (!transactionMeta) {
-    throw new Error(
-      `publishBatchSmartTransactionHook: Could not find transaction with id ${lastTransaction.id}`,
-    );
-  }
-
-  const { isSmartTransaction, featureFlags, isHardwareWalletAccount } =
-    getSmartTransactionCommonParams(flatState, transactionMeta.chainId);
-
-  if (!isSmartTransaction) {
-    return undefined;
-  }
-
-  return submitBatchSmartTransactionHook({
-    transactions,
-    transactionController,
-    smartTransactionsController,
-    controllerMessenger: hookControllerMessenger,
-    isSmartTransaction,
-    isHardwareWallet: isHardwareWalletAccount,
-    // @ts-expect-error Smart transaction selector return type does not match FeatureFlags type from hook
-    featureFlags,
-    transactionMeta,
-  });
 }
